+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * `unnest()` weakens test of "atomicity" to restore previous behaviour when
   unnesting factors and dates (#407).
+  
+* `separate()` now accepts `NA` as a column name in the `into` argument to
+  denote columns which are omitted from the result. (@markdly, #397).
 
 # tidyr 0.8.0